move icon helpers into an app-icon screen and out of miscellany
